+++ src/main/java/com/yuuko/core/Configuration.java
@@ -8,7 +8,7 @@ import java.io.FileReader;
 import java.util.HashMap;
 
 public class Configuration {
-    public static final String VERSION = "13-01-2019_1";
+    public static final String VERSION = "15-01-2019_1";
     public static String AUTHOR;
     public static String AUTHOR_WEBSITE;
     public static String SUPPORT_GUILD;